+++ src/migrations/m260309_145000_release_0105_noop.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Klick\Agents\migrations;
+
+use craft\db\Migration;
+
+class m260309_145000_release_0105_noop extends Migration
+{
+    public function safeUp(): bool
+    {
+        return true;
+    }
+
+    public function safeDown(): bool
+    {
+        return true;
+    }
+}